Agent2/requirements.txt
flask==2.3.3
mysql-connector-python==8.1.0
google-generativeai==0.3.2
python-dotenv==1.0.0
pandas>=2.1.0
sqlalchemy